Add missing exception handling for PyString_FromString() calls
In Python 3 these will all involve UTF-8 decoding, which can easily trigger
exceptions for invalid UTF-8 data. For instance this happened when running
the tests and a commit was made with the “svn:log” string accidentally not
terminated. The string had random non UTF-8 data appended to it.

Patch by Martin Panter.