Fix raising deferred socket exception
Python 3 deletes the exception variable at the end of the exception handler,
so we have to explicitly save it in another variable. Make sure that variable
is eventually deleted to avoid the traceback causing a reference cycle.

Verified by running the code

subvertpy.ra_svn.SVNClient("svn://wbr")

where “wbr” is in my “/etc/hosts” file but is not connectable.

This also breaks an import dependency cycle between “subvertpy.ra” and
“subvertpy.ra_svn”, because Python 3 is less forgiving than Python 2.

Patch by Martin Panter

Signed-off-by: Jelmer Vernooij <jelmer@samba.org>