Pass absolute path to svn_client_info routine.
Save and pass the path prefix so that we can
recover the relative path in the output.

Test fails, but this time I believe that the test is wrong.
Changed to expect the behaviour of the C-api. Note this
is different if there is a working copy only operation
or one that involves the remote repo.