In test_has_binary_prop() should assertTrue, not False.
The file is binary so the has_binary_prop method should
return True.

It does indeed do this on 1.8, but fails with the test as
written on 1.6.

On 1.6 it appears there is a limitation that requires that
the WorkingCopy() is opened the same kind (absolute or relative)
of path that is used in the call to prop_get() to make
this work.