Reverse result of test_translated_stream and limit to 1.7+
The test as written fails on 1.8 however I don't believe that
the test is correct.  I believe that the Normal Form is the repository
format and the other format is the Working Copy format.  So the
the normal form has keywords contracted.  This is the observed
behaviour on 1.8.

So change the test to check for un-expanded form and limit to
greater than version >= 1.7