Change two tests to work with previously committed revision.
The tests test_process_committed_queue and test_transmit_text_deltas
fail with an assertion on 1.8. I think that they require a previously
committed revision (or a copy of a previously committed file to
a new name).

Changed to commit a version first and adjusted as necessary to
check for version 2 rather than 1.